+++ app/Domain/Pipelines/InvalidPipelineTemplateException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Domain\Pipelines;
+
+use InvalidArgumentException;
+
+final class InvalidPipelineTemplateException extends InvalidArgumentException
+{
+}